+++ resources/views/includes/debug.blade.php
@@ -1,0 +1,5 @@
+<div>
+    @if ($dumpFilters)
+        @dump($filters)
+    @endif
+</div>